feat(validation): crack down on more validation
right now the size and type fields were being set via the data sent in the clients api request, however this needs to be automatically set for security sake, and should not allow being sent.

add validation to ensure the data sent to the api endpoint is null upon receipt, and that the data is properly handled within the service layer to calculate those values automatically

+++ src/main/java/com/phoenixware/inventorynexus/shared/entity/Asset.java
@@ -36,7 +36,7 @@ public class Asset {
     private String alt;
 
     @Column(name = "size")
-    private Integer size;
+    private Long size;
 
     @Column(name = "type")
     private String type;
